Added SQLiteConnection.Initialize(), required to Android
In initialization it tells path of temporary directory to libsqlite.so.
This is required to prevent `CannotOpen` error when libsqlite.so
attempts to create a temporary file.

The error comes that libsqlite.so fails to find out temp directory
since Android has no default temp directory.

In addition, there seems no way to know the path to temporary directory
from NDK libary throw JNI.

Even though Unity provides the way to get to know the path, it seems
unstable; it may get the path in a app, but not in other app - it may
depend on app environment.

So it should be safer way to tell the temp directory from C# side.

On iOS or Unity Editor it does nothing since it works fine without the
initialization.

- `SQLiteConnection.Initialize()` is called implicitly within
   `static SQLiteConnection()`
- `SQLiteConnection.Initialize()` should be called in the MainThread.
  So it may be required to call `SQLiteConnection.Initialize()`
  explicitly ahead of accessing SQLite storage from worker threads.